record.receivers: refactor enhance receiver
Refactor to explicitly show what 'populator' methods are running on
each data type.
